Allow directory scanning of input files
